test: HttpRequestUtilsTest 를 프로덕션 코드에 맞게 수정한다.
